# common/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /v1/article/acceptAnswer:
        post:
            tags:
                - ContentArticleService
            description: 采纳评论
            operationId: ContentArticleService_AcceptAnswer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.AcceptAnswerArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.AcceptAnswerArticleReply'
    /v1/article/add:
        post:
            tags:
                - ContentArticleService
            description: 新增文章
            operationId: ContentArticleService_Add
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.AddArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.AddArticleReply'
    /v1/article/addPostscript:
        post:
            tags:
                - ContentArticleService
            description: 添加附言
            operationId: ContentArticleService_AddPostscript
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.AddPostscriptArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.AddPostscriptArticleReply'
    /v1/article/collect:
        post:
            tags:
                - ContentArticleService
            description: 收藏文章
            operationId: ContentArticleService_Collect
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.CollectArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.CollectArticleReply'
    /v1/article/delete:
        post:
            tags:
                - ContentArticleService
            description: 删除文章（仅草稿）
            operationId: ContentArticleService_Delete
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.DeleteArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.DeleteArticleReply'
    /v1/article/getOne:
        post:
            tags:
                - ContentArticleService
            description: 查询单篇文章
            operationId: ContentArticleService_GetOne
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.GetArticleOneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.GetArticleOneReply'
    /v1/article/like:
        post:
            tags:
                - ContentArticleService
            description: 点赞文章
            operationId: ContentArticleService_Like
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.LikeArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.LikeArticleReply'
    /v1/article/page:
        post:
            tags:
                - ContentArticleService
            description: 查询文章
            operationId: ContentArticleService_Page
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.PageArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.PageArticleReply'
    /v1/article/publish:
        post:
            tags:
                - ContentArticleService
            description: 发布文章（从草稿发布）
            operationId: ContentArticleService_Publish
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.PublishArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.PublishArticleReply'
    /v1/article/reward:
        post:
            tags:
                - ContentArticleService
            description: 打赏文章
            operationId: ContentArticleService_Reward
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.RewardArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.RewardArticleReply'
    /v1/article/thank:
        post:
            tags:
                - ContentArticleService
            description: 感谢文章
            operationId: ContentArticleService_Thank
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.ThankArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.ThankArticleReply'
    /v1/article/update:
        post:
            tags:
                - ContentArticleService
            description: 修改文章（管理员使用）
            operationId: ContentArticleService_Update
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.UpdateArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.UpdateArticleReply'
    /v1/article/updateDraft:
        post:
            tags:
                - ContentArticleService
            description: 更新草稿
            operationId: ContentArticleService_UpdateDraft
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.UpdateArticleDraftRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.UpdateArticleDraftReply'
    /v1/article/watch:
        post:
            tags:
                - ContentArticleService
            description: 关注文章
            operationId: ContentArticleService_Watch
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.content.v1.WatchArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.content.v1.WatchArticleReply'
...
components:
    schemas:
...
        common.api.content.v1.AcceptAnswerArticleReply:
            type: object
            properties: {}
        common.api.content.v1.AcceptAnswerArticleRequest:
            type: object
            properties:
                article_id:
                    type: string
                    description: 文章Id
                comment_id:
                    type: string
                    description: 评论Id
        common.api.content.v1.AddArticleReply:
            type: object
            properties:
                article:
                    allOf:
                        - $ref: '#/components/schemas/common.api.content.v1.Article'
                    description: 文章
        common.api.content.v1.AddArticleRequest:
            type: object
            properties:
                article:
                    allOf:
                        - $ref: '#/components/schemas/common.api.content.v1.ArticleSave'
                    description: 文章
...
        common.api.content.v1.AddPostscriptArticleReply:
            type: object
            properties:
                article_postscript:
                    allOf:
                        - $ref: '#/components/schemas/common.api.content.v1.ArticlePostscript'
                    description: 附言
        common.api.content.v1.AddPostscriptArticleRequest:
            type: object
            properties:
                article_id:
                    type: string
                    description: 文章Id
                content:
                    type: string
                    description: 内容
...
        common.api.content.v1.CollectArticleReply:
            type: object
            properties: {}
        common.api.content.v1.CollectArticleRequest:
            type: object
            properties:
                article_id:
                    type: string
                    description: 文章Id
                active:
                    type: boolean
                    description: 是否收藏
...
        common.api.content.v1.DeleteArticleReply:
            type: object
            properties: {}
        common.api.content.v1.DeleteArticleRequest:
            type: object
            properties:
                article_id:
                    type: string
                    description: 文章Id
...